[Translation] Use trait for translatable message parameters
