+++ src/Symfony/Component/Translation/TranslatableMessage.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Component\Translation;
 
+use Symfony\Component\Translation\Message\TranslatableParametersTrait;
 use Symfony\Contracts\Translation\TranslatableInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
@@ -19,8 +20,9 @@ use Symfony\Contracts\Translation\TranslatorInterface;
  */
 class TranslatableMessage implements TranslatableInterface
 {
+    use TranslatableParametersTrait;
+
     private $message;
-    private $parameters;
     private $domain;
 
     public function __construct(string $message, array $parameters = [], string $domain = null)
@@ -40,11 +42,6 @@ class TranslatableMessage implements TranslatableInterface
         return $this->message;
     }
 
-    public function getParameters(): array
-    {
-        return $this->parameters;
-    }
-
     public function getDomain(): ?string
     {
         return $this->domain;
@@ -52,11 +49,11 @@ class TranslatableMessage implements TranslatableInterface
 
     public function trans(TranslatorInterface $translator, string $locale = null): string
     {
-        return $translator->trans($this->getMessage(), array_map(
-            static function ($parameter) use ($translator, $locale) {
-                return $parameter instanceof TranslatableInterface ? $parameter->trans($translator, $locale) : $parameter;
-            },
-            $this->getParameters()
-        ), $this->getDomain(), $locale);
+        return $translator->trans(
+            $this->getMessage(),
+            $this->getTranslatedParameters($translator, $locale),
+            $this->getDomain(),
+            $locale
+        );
     }
 }
